requirements.txt
aiogram~=3.13
python-dotenv~=1.0
aiosqlite~=0.19
pydantic~=2.5
markdown2~=2.4
pytest~=7.4
pytest-asyncio~=0.23
pendulum~=2.1
aiofiles~=23.2
python-multipart~=0.0